% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_networks.R
\name{add_network_id}
\alias{add_network_id}
\title{Add a network_id column to identifiers_df}
\usage{
add_network_id(identifiers_df = load_identifiers_df(),
  temporary_files = NULL, continue_from_temporary = FALSE)
}
\arguments{
\item{identifiers_df}{A data frame, typically created with extract extract_identifiers() and loaded with load_identifiers_df().}
}
\value{
A data.frame (a tibble) including a network_id column grouping all domains that have elements in common.
}
\description{
Add a network_id column to identifiers_df
}
\examples{

}
