% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_data.R
\name{archive_screenshots}
\alias{archive_screenshots}
\title{Archives screenshots}
\usage{
archive_screenshots(language = NULL)
}
\description{
It archives  folders in .tar.gz files, and stores them in a dated sub-folder within the `archive` sub-folder.
}
