% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_exists}
\alias{check_if_exists}
\title{Check if corresponding information for a given domain exists for a given data type and a given time period}
\usage{
check_if_exists(domain, type = "homepage", language = NULL,
  since = Sys.Date() - 31, simplify = TRUE)
}
\arguments{
\item{domain}{A domain name (e.g. "example.com").}

\item{type}{Type of file to check, defaults to "homepage". Alternative options: "screenshots".}
}
\value{
Nothing, used for its side effects.
}
\description{
Check if corresponding information for a given domain exists for a given data type and a given time period
}
\examples{

}
