% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_excluded_fb_admins}
\alias{default_excluded_fb_admins}
\title{fb_admins ID to be excluded by default}
\format{
A character vector.
}
\usage{
default_excluded_fb_admins
}
\description{
This includes ID such as "YOUR USER ID" left by lazy developers.
}
\keyword{datasets}
