% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_excluded_fb_app_id}
\alias{default_excluded_fb_app_id}
\title{fb_app_id ID to be excluded by default}
\format{A character vector.}
\usage{
default_excluded_fb_app_id

default_excluded_fb_app_id
}
\description{
This includes ID such as "Facebook App ID here" or the default app id (966242223397117) left by lazy developers.

This includes ID such as "Facebook App ID here" or the default app id (966242223397117) left by lazy developers.
}
\keyword{datasets}
