% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_excluded_ua}
\alias{default_excluded_ua}
\title{ua ID to be excluded by default}
\format{A character vector.}
\usage{
default_excluded_ua
}
\description{
This includes ID such as the ua id used by large hosts such as wordpress.com.
}
\keyword{datasets}
