% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_identifiers}
\alias{default_identifiers}
\title{Type of snippets used to identify networked websites}
\format{
A character vector.
}
\usage{
default_identifiers
}
\description{
This includes, among others, Google Analytics and Google Adsense codes, as well as Facebook identifiers
}
\keyword{datasets}
