% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_domains.R
\name{extract_domains_from_tweets}
\alias{extract_domains_from_tweets}
\title{Get screenshots and place them in relevant subfolder}
\usage{
extract_domains_from_tweets(language = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Get screenshots and place them in relevant subfolder
}
\examples{

}
