% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_keywords.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract most frequently used words from news titles}
\usage{
extract_keywords(languages = c("ar", "bg", "cs", "da", "de", "el", "en",
  "es", "et", "fi", "fr", "hr", "hu", "it", "lt", "lv", "nl", "pl", "pt",
  "ro", "ru", "sk", "sl", "sv", "sw", "tr", "zh"), n = 10,
  date = Sys.Date(), store = TRUE)
}
\arguments{
\item{n}{An integer. Number of words to keep.}

\item{date}{Only news downloaded in the given date will be considered. Defaults to current day. To get data for the previous day, use `Sys.Date()-1`}

\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Extract most frequently used words from news titles
}
\examples{

}
