% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_networks.R
\name{find_related_domains}
\alias{find_related_domains}
\title{Get screenshots and place them in relevant subfolder}
\usage{
find_related_domains(domain, identifiers_df,
  identifiers = default_identifiers, language = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Get screenshots and place them in relevant subfolder
}
\examples{

}
