% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_websites.R
\name{get_homepage}
\alias{get_homepage}
\title{Download homepages by url in relevant sub-folder.}
\usage{
get_homepage(domain = NULL, language = NULL)
}
\arguments{
\item{domain}{A character vector of one or more domain names.}
}
\value{
Nothing, used for its side effects.
}
\description{
Download homepages by url in relevant sub-folder, divide by date of download. By default, does not overwrite, and shuffles the order in which given domain names are downloaded.
}
\examples{

}
