% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_website_data.R
\name{get_screenshot}
\alias{get_screenshot}
\title{Get screenshots and place them in relevant subfolder}
\usage{
get_screenshot(url, width = 1280, height = 1280, language = NULL)
}
\arguments{
\item{url}{A character vector of one or more urls.}
}
\value{
Nothing, used for its side effects.
}
\description{
Get screenshots and place them in relevant subfolder
}
\examples{

}
