% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Searches keywords on Twitter.}
\usage{
get_tweets(keywords = NULL, date = Sys.Date(), language,
  n_tweets = 1000, wait = 10, store = TRUE)
}
\arguments{
\item{keywords}{Defaults to NULL. If NULL, tries to import previously stored keywords for given language and date. If a data frame, it assumes that the first columns includes the keywords. If a character vector, it processes it directly.}

\item{date}{Used to find keywords locally, if none are provided. Defaults to current day. To get data for the previous day, use `Sys.Date()-1`}

\item{n_tweets}{An integer. Number of tweets to request.}
}
\value{
A data.frame (a tibble) with `n` number of rows and two columns, `words` and `n` for number of occurrences.
}
\description{
It assumes that a valid token is already present in the working environment. For details, see https://rtweet.info/articles/auth.html
}
\examples{

}
