% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_networks.R
\name{nwd_add_network_id}
\alias{nwd_add_network_id}
\title{Add a network_id column to identifiers_df}
\usage{
nwd_add_network_id(
  identifiers_df = nwd_load_identifiers_df(),
  language = NULL,
  identifiers = NULL,
  temporary_files = NULL,
  continue_from_temporary = FALSE
)
}
\arguments{
\item{identifiers_df}{A data frame, typically created with extract extract_identifiers() and loaded with load_identifiers_df().}

\item{temporary_files}{Defaults to NULL. If given, it should be an integer corresponding to the number of temporary files to create while processing the data. E.g. If if there are 100.000 rows to process, and temporary_files is set to 100, a total of 100 files will be stored in the temporary folder, one each 1.000 rows.}
}
\value{
A data.frame (a tibble) including a network_id column grouping all domains that have elements in common.
}
\description{
Add a network_id column to identifiers_df
}
\examples{

}
