% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_adjust_folder_names}
\alias{nwd_adjust_folder_names}
\title{Adjusts file names of folders with single digits}
\usage{
nwd_adjust_folder_names(folder = "tweets", language = NULL,
  date = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Adjusts file names of folders with single digits
}
\examples{

}
