% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_adjust_folder_names_nested}
\alias{nwd_adjust_folder_names_nested}
\title{Adjusts names of nested date folders}
\usage{
nwd_adjust_folder_names_nested(folder = "tweets", language = NULL, date = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Adjusts names of nested date folders
}
\examples{

}
