% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_archive}
\alias{nwd_archive}
\title{Archive data in compressed files}
\usage{
nwd_archive(date = Sys.Date(), folder = "tweets",
  timeframe = "monthly", language = "it", filetype = "rds",
  to_googledrive = FALSE)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Archive data in compressed files
}
\examples{

}
