% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_backup_to_googledrive}
\alias{nwd_backup_to_googledrive}
\title{Backup archived files to googledrive}
\usage{
nwd_backup_to_googledrive(date = NULL, folder = "tweets",
  timeframe = "daily", language = NULL, filetype = "rds",
  show_filenames_only = FALSE)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Backup archived files to googledrive
}
\examples{

}
