% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nwd_check_if_exists}
\alias{nwd_check_if_exists}
\title{Check if corresponding information for a given domain exists for a given data type and a given time period}
\usage{
nwd_check_if_exists(
  domain,
  type = "homepage",
  language = NULL,
  since = Sys.Date() - 91,
  simplify = TRUE
)
}
\arguments{
\item{domain}{A domain name (e.g. "example.com").}

\item{type}{Type of file to check, defaults to "homepage". Curent alternative options: "screenshots" and "ads".}

\item{since}{A date. `nwd_check_if_exists()` returns TRUE only if the given link has been downloaded or (failed to download) after the given date.}
}
\value{
If `simplify` is TRUE, it returns a logical vector of length 1, either TRUE or FALSE. If `simplify == FALSE`, it returns a data frame with details on the availability of files related to the given domain.
}
\description{
Check if corresponding information for a given domain exists for a given data type and a given time period
}
\examples{

}
