% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_networks.R
\name{nwd_create_domain_graph_identifiers}
\alias{nwd_create_domain_graph_identifiers}
\title{Create a graph of the network of one or more domains}
\usage{
nwd_create_domain_graph_identifiers(network,
  identifiers_df = nwd_load_identifiers_df(), identifiers = NULL,
  only_shared_identifiers = TRUE, plot = TRUE)
}
\arguments{
\item{identifiers_df}{A data frame, typically created with extract extract_identifiers() and loaded with load_identifiers_df().}
}
\value{
A data.frame (a tibble) including a network_id column grouping all domains that have elements in common.
}
\description{
Create a graph of the network of one or more domains
}
\examples{

}
