% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_download_from_googledrive}
\alias{nwd_download_from_googledrive}
\title{Download files archived in Google Drive}
\usage{
nwd_download_from_googledrive(date = NULL, folder = "tweets",
  timeframe = "daily", language = NULL, filetype = "rds",
  overwrite = FALSE)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Download files archived in Google Drive
}
\examples{

}
