% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_urls_from_tweets.R
\name{nwd_expand_urls_from_tweets}
\alias{nwd_expand_urls_from_tweets}
\title{Extract links from tweets and expands shortened urls}
\usage{
nwd_expand_urls_from_tweets(tweets, n_char = 30, n_retry = 1)
}
\arguments{
\item{tweets}{A data frame of tweets, as created by `rtweet`.}

\item{n_char}{An integer, defaults to 30. Number of characters in the url: only urls shorter than n_char will be expanded, others will be kept as they are.}

\item{n_retry}{An integer, defaults to 1. Number of times it tries to expand urls if first attempt fails. Set to 0 for attempting only once.}
}
\value{
A data.frame (a tibble) with 9 columns: "user_id", "status_id", "created_at", "screen_name", text", "orig_url","expanded_url", "status_code", "url"
}
\description{
Extract links from tweets and expands shortened urls
}
\examples{

}
