% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_domains.R
\name{nwd_extract_domains}
\alias{nwd_extract_domains}
\title{Extract domain names from pre-processed tweets}
\usage{
nwd_extract_domains(language = NULL, date = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}

\item{date}{Only news downloaded in the given date will be considered. Defaults to current day. To get data for the previous day, use `Sys.Date()-1`}
}
\value{
A data.frame (a tibble) with one column corresponding to the domain names found on a given date.
}
\description{
Extract domain names from pre-processed tweets
}
\examples{

}
