% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_domains.R
\name{nwd_extract_domains_from_tweets}
\alias{nwd_extract_domains_from_tweets}
\title{Extract domain names from tweets}
\usage{
nwd_extract_domains_from_tweets(language = NULL)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Extract domain names from tweets
}
\examples{

}
