% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_identifiers.R
\name{nwd_extract_identifiers_from_backup}
\alias{nwd_extract_identifiers_from_backup}
\title{Extract identifiers from backuped homepages}
\usage{
nwd_extract_identifiers_from_backup(language)
}
\arguments{
\item{language}{A character vector of length one corresponding to a language two-letter code.}
}
\value{
Nothing, used for its side effects.
}
\description{
Extract identifiers from backuped homepages
}
\examples{

}
