% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ads.R
\name{nwd_get_ads}
\alias{nwd_get_ads}
\title{Download ads.txt in relevant sub-folder.}
\usage{
nwd_get_ads(
  domain = NULL,
  language = NULL,
  shuffle = TRUE,
  since = Sys.Date() - 91
)
}
\arguments{
\item{domain}{A character vector of one or more domain names.}

\item{since}{A date. Passed to `nwd_check_if_exists()`. Only domains that have not been downloaded since the given date are downloaded.}
}
\value{
Nothing, used for its side effects.
}
\description{
Download ads.txt by url in relevant sub-folder, divide by date of download. By default, does not overwrite, and shuffles the order in which given domain names are downloaded.
}
\examples{

}
