% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{nwd_get_tweets}
\alias{nwd_get_tweets}
\title{Searches keywords on Twitter.}
\usage{
nwd_get_tweets(
  keywords = NULL,
  date = NULL,
  language = NULL,
  n_keywords = 10,
  n_tweets = 1000,
  tweet_type = "recent",
  wait = 10,
  store = TRUE
)
}
\arguments{
\item{keywords}{A character vector or a data frame, with keywords in the first column. Defaults to NULL. If NULL, tries to import previously stored keywords for given language and date. If a data frame, it assumes that the first columns includes the keywords. If a character vector, it processes it directly.}

\item{date}{Used to find keywords locally, if none are provided. Defaults to latest available. To systematically get data for the previous day, use `Sys.Date()-1`}

\item{n_keywords}{An integer, defaults to NULL.}

\item{n_tweets}{An integer. Number of tweets to request.}

\item{tweet_type}{Passed to `rtweet`. Defaults to "recent". Other valid types include "mixed" and "popular".}
}
\value{
A data.frame (a tibble) with `n` number of rows and two columns, `words` and `n` for number of occurrences.
}
\description{
It assumes that a valid token is already present in the working environment. For details, see https://rtweet.info/articles/auth.html
}
\examples{

}
