% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nwd_list_available_backups}
\alias{nwd_list_available_backups}
\title{List available backups}
\usage{
nwd_list_available_backups(date = NULL, folder = "tweets",
  timeframe = "daily", language = NULL, filetype = "rds")
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
A tibble (a data frame) with three columns: name, date, location.
}
\description{
List available backups
}
\examples{

}
