% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_websites.R
\name{download_homepage}
\alias{download_homepage}
\title{Download homepages by url in relevant sub-folder.}
\usage{
download_homepage(domain, language)
}
\arguments{
\item{url}{A character vector of one or more domain names.}
}
\value{
Nothing, used for its side effects.
}
\description{
Download homepages by url in relevant sub-folder, divide by date of download. By default, does not overwrite.
}
\examples{

}
