% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_id.R
\name{extract_identifiers}
\alias{extract_identifiers}
\title{Extract identifiers from web pages}
\usage{
extract_identifiers(language = NULL, progress_bar = TRUE)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Extract identifiers from web pages
}
\examples{

}
