% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_website_data.R
\name{get_screenshot}
\alias{get_screenshot}
\title{Get screenshots and place them in relevant subfolder}
\usage{
get_screenshot(
  domain,
  width = 1280,
  height = 1280,
  language = NULL,
  shuffle = TRUE
)
}
\arguments{
\item{domain}{A character vector of one or more domains (e.g. "example.com").}

\item{shuffle}{Logical, defaults to TRUE. If TRUE, order in which domains are processed is randomised.}
}
\value{
Nothing, used for its side effects.
}
\description{
Get screenshots and place them in relevant subfolder
}
\examples{

}
