% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_archive}
\alias{nwd_archive}
\title{Archive data in compressed files}
\usage{
nwd_archive(date = NULL, folder = "tweets", timeframe = "daily",
  language = NULL, filetype = "rds")
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Archive data in compressed files
}
\examples{

}
