% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nwd_load_latest_identifiers_df}
\alias{nwd_load_latest_identifiers_df}
\title{Load latest identifiers_df}
\usage{
nwd_load_latest_identifiers_df(language = NULL)
}
\value{
A data.frame (a tibble), typically generated with `extract_identifiers()`
}
\description{
Load latest identifiers_df
}
\examples{

}
