% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{nwd_restore}
\alias{nwd_restore}
\title{Restore data from compressed files}
\usage{
nwd_restore(date = NULL, folder = "tweets", timeframe = "daily",
  language = NULL, filetype = "rds", filenames_only = FALSE)
}
\arguments{
\item{language}{A character vector of language two letter codes. Defaults to NULL. If NULL, processes available languages.}
}
\value{
Nothing, used for its side effects.
}
\description{
Restore data from compressed files
}
\examples{

}
